# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CVE API",
    "version": "1.0.0",
    "description": "Simple CVE listing and search API"
  },
  "paths": {
    "/list": {
      "get": {
        "summary": "List CVEs with optional filters",
        "parameters": [
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "sort", "in": "query", "schema": { "type": "string", "enum": ["published","score"] }, "description": "Sort by published date or score" },
          { "name": "minScore", "in": "query", "schema": { "type": "number" }, "description": "Minimum score (inclusive)" },
          { "name": "maxScore", "in": "query", "schema": { "type": "number" }, "description": "Maximum score (inclusive)" },
          { "name": "year", "in": "query", "schema": { "type": "integer" }, "description": "Published year to filter by" },
          { "name": "scoreVersion", "in": "query", "schema": { "type": "string", "enum": ["v3.1","v4.0","v3.0","v2.0","effective"] }, "description": "Which metric to use for score-based filtering/sorting" }
        ],
        "responses": {
          "200": {
            "description": "Array of CVE records",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/CVERecord" } }
              }
            }
          }
        }
      }
    },
    "/findID": {
      "get": {
        "summary": "Find CVE by ID",
        "parameters": [
          { "name": "search", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Array of matching CVE records",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/CVERecord" } } } }
          }
        }
      }
    },
    "/findText": {
      "get": {
        "summary": "Full-text search across CVE fields",
        "parameters": [
          { "name": "search", "in": "query", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Array of matching CVE records",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/CVERecord" } } } }
          }
        }
      }
    }
    ,"/index/fields": {
      "get": {
        "summary": "List Bleve indexed field names",
        "responses": {
          "200": {
            "description": "Array of field name strings",
            "content": { "application/json": { "schema": { "type": "array", "items": { "type": "string" } } } }
          }
        }
      }
    },
    "/index/mappings": {
      "get": {
        "summary": "Return the Bleve index mapping as JSON",
        "responses": {
          "200": {
            "description": "Bleve index mapping JSON",
            "content": { "application/json": { "schema": { "type": "object" } } }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CVERecord": {
        "type": "object",
        "properties": {
          "dataType": { "type": "string" },
          "dataVersion": { "type": "string" },
          "cveMetadata": {
            "type": "object",
            "properties": {
              "cveId": { "type": "string" },
              "datePublished": { "type": "string", "format": "date-time" }
            }
          },
          "containers": {
            "type": "object",
            "properties": {
              "cna": {
                "type": "object",
                "properties": {
                  "title": { "type": "string" },
                  "descriptions": { "type": "array", "items": { "type": "object", "properties": { "lang": { "type": "string" }, "value": { "type": "string" } } } }
                }
              }
            }
          }
        }
      }
    }
  }
}